requirements.txt
PySide6~=6.7.2
PyPDFForm~=1.4.34
python-dotenv~=1.0.1
supabase~=2.10.0
bcrypt~=4.2.0
numpy~=2.1.3
pillow~=11.0.0